move sso command to --sso on the login command
